REFACTOR AND RESTYLE
