+++ projects/client/src/components/AdminHeaderPageLayout.jsx
@@ -7,6 +7,20 @@ function AdminHeaderPageLayout(props) {
     return forPage === "warehouse";
   };
 
+  const RenderForWarehouseMgtPage = () => {
+    return (
+      <div className="grid grid-cols-2 row-span-1 row-start-4 gap-20 lg:grid-cols-4 lg:gap-8">
+        <button
+          onClick={() => setFirstSubPage(true)}
+          className="py-1 px-1 bg-primary text-white text-sm 
+      font-semibold hover:bg-slate-800 lg:text-lg"
+        >
+          Warehouse <i className="uil uil-plus"></i>
+        </button>
+      </div>
+    );
+  };
+
   return (
     <div className="w-full text-center row-span-2 grid grid-rows-4">
       <h1
@@ -16,15 +30,7 @@ function AdminHeaderPageLayout(props) {
         {firstSubPageClicked ? headerTitle[0] : headerTitle[1]} {forPage === "report" ? "" : "Management"}
       </h1>
       {isItForWarehousePage() ? (
-        <div className="grid grid-cols-2 row-span-1 row-start-4 gap-20 lg:grid-cols-4 lg:gap-8">
-          <button
-            onClick={() => setFirstSubPage(true)}
-            className="py-1 px-1 bg-primary text-white text-sm 
-            font-semibold hover:bg-slate-800 lg:text-lg"
-          >
-            Warehouse <i className="uil uil-plus"></i>
-          </button>
-        </div>
+        <RenderForWarehouseMgtPage />
       ) : (
         <div className="grid grid-cols-2 row-start-4 row-span-1 gap-20 lg:grid-cols-4 lg:gap-8">
           <button
@@ -33,8 +39,9 @@ function AdminHeaderPageLayout(props) {
               setFirstSubPage(true);
               setSecondSubPage(false);
             }}
-            className="py-1 px-1 bg-primary text-white text-sm font-semibold hover:bg-slate-800
-           lg:text-lg md:text-sm  disabled:bg-slate-100 disabled:text-slate-400 disabled:hover:cursor-not-allowed"
+            className="py-1 px-1 text-sm bg-slate-100 text-slate-400  
+            font-semibold hover:bg-slate-800 hover:text-white
+           lg:text-lg md:text-sm disabled:hover:cursor-not-allowed disabled:hover:bg-slate-100 disabled:hover:text-slate-400"
           >
             {headerTitle[0]}
           </button>
@@ -44,9 +51,9 @@ function AdminHeaderPageLayout(props) {
               setFirstSubPage(false);
               setSecondSubPage(true);
             }}
-            className="py-1 px-1 bg-white text-primary text-sm 
-          font-semibold border-primary border-2 hover:bg-slate-800 disabled:hover:cursor-not-allowed
-          hover:text-white disabled:bg-slate-100 disabled:text-slate-400 disabled:border-primaryLight lg:text-lg"
+            className="py-1 px-1 text-sm bg-slate-100 text-slate-400 
+          font-semibold hover:bg-slate-800 disabled:hover:cursor-not-allowed
+          hover:text-white lg:text-lg disabled:hover:bg-slate-100 disabled:hover:text-slate-400"
           >
             {headerTitle[1]}
           </button>
