+++ projects/client/src/feature/admin_product/component/manage_stock/delete_data/DeleteModal.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import RenderWarehouseOption from "../RenderWarehouseOption";
-import { deleteStock, getProductsStocks, getStock } from "../../../";
+import { deleteStock, getStock } from "../../../";
 import ClosedBtnModal from "../../../../../components/ClosedBtnModal";
 
 function DeleteModal(props) {
