+++ projects/client/src/feature/stock_mutation/component/stock_mutation/StockMutationBody.jsx
@@ -1,62 +1,34 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import AdminPagination from "../../../../components/AdminPagination";
 import RenderBodyData from "../../../../components/RenderBodyData";
 import AddModal from "./add_request/AddModal";
 import SelectFilter from "../../../../components/SelectFilter";
 import RenderWarehouse from "../../../admin/component/all_admin/edit_data/RenderWarehouse";
-import { getMutationRequests } from "../../";
 import RenderMutationData from "../RenderMutationData";
 import NoData from "../../../../components/NoData";
 import UpdateModal from "./update_mutation/UpdateModal";
+import { useStockMutationBody } from "../../util/useStockMutationBody";
 
 function StockMutationBody(props) {
   const { admin, warehouses } = props;
-  const [createNewRequest, setNewRequest] = useState(false);
-  const [pageNum, setPageNum] = useState(1);
-  const [totalPage, setTotalPage] = useState(1);
-  const [filterState, setFilterState] = useState({
-    id_warehouse: admin.id_warehouse ? admin.id_warehouse : "",
-    mutationType: "",
-    status: "",
-  });
-  const [mutationList, setMutationList] = useState([]);
-  const [singleItemClicked, setSingleItemClicked] = useState(false);
-  const [singleData, setSingleData] = useState({});
-  const OFFSET = 7;
-  const LIMIT = 7;
-
-  useEffect(() => {
-    (async () => {
-      fetchingData();
-    })();
-  }, [filterState, pageNum]);
-
-  const fetchingData = async () => {
-    const response = await getMutationRequests(OFFSET, LIMIT, pageNum, filterState);
-    setMutationList([...response?.result.dataToSend]);
-    setTotalPage(response.result.totalPage);
-  };
-
-  const warehouseOnChange = async (e) => {
-    setFilterState((prevState) => ({ ...prevState, id_warehouse: e.target.value }));
-  };
-
-  const mutationOnChange = async (e) => {
-    setFilterState((prevState) => ({ ...prevState, mutationType: e.target.value }));
-  };
-
-  const statusOnChange = async (e) => {
-    setFilterState((prevState) => ({ ...prevState, status: e.target.value }));
-  };
-
-  const singleItemClickedHandler = (singleData) => {
-    setSingleItemClicked(true);
-    setSingleData({ ...singleData });
-  };
-
-  const isWarehouseFilterDisabled = () => {
-    return admin.id_role !== 1;
-  };
+  const {
+    createNewRequest,
+    setNewRequest,
+    pageNum,
+    setPageNum,
+    totalPage,
+    filterState,
+    mutationList,
+    singleItemClicked,
+    setSingleItemClicked,
+    singleData,
+    warehouseOnChange,
+    mutationOnChange,
+    statusOnChange,
+    singleItemClickedHandler,
+    isWarehouseFilterDisabled,
+    fetchingData,
+  } = useStockMutationBody(admin);
 
   return (
     <>
